created form component for search, updated api for next build, tweaked translations, components
